requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu117
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
ffmpeg-python==0.2.0
gunicorn==20.1.0
requests==2.28.2
openai==0.27.2
openai-whisper==20230314
python-dotenv==1.0.0
tiktoken==0.3.1
torch
torchvision
torchaudio